+++ src/suggesters/FolderSuggester.ts
@@ -1,10 +1,14 @@
-// Credits go to Liam's Periodic Notes Plugin: https://github.com/liamcain/obsidian-periodic-notes
-
-import { TAbstractFile, TFolder } from "obsidian";
-import { TextInputSuggest } from "./suggest";
+import { AbstractInputSuggest, TAbstractFile, TFolder } from "obsidian";
 import { Iadapter } from "src/dataStore/adapter";
 
-export class FolderSuggest extends TextInputSuggest<TFolder> {
+export class FolderSuggest extends AbstractInputSuggest<TFolder> {
+    private readonly inputEl: HTMLInputElement;
+
+    constructor(inputEl: HTMLInputElement) {
+        super(Iadapter.instance.app, inputEl);
+        this.inputEl = inputEl;
+    }
+
     getSuggestions(inputStr: string): TFolder[] {
         const abstractFiles = Iadapter.instance.vault.getAllLoadedFiles();
         const folders: TFolder[] = [];
@@ -27,6 +31,7 @@ export class FolderSuggest extends TextInputSuggest<TFolder> {
     }
 
     selectSuggestion(file: TFolder): void {
+        this.setValue(file.path);
         this.inputEl.value = file.path;
         this.inputEl.trigger("input");
         this.close();